remove useless imports, SignIn=>Prompt, SignInController=>SignIn
