+++ packages/frontend/src/components/CreateAccount/index.jsx
@@ -4,7 +4,6 @@ import {
   Button,
   CssBaseline,
   FormControl,
-  Input,
   InputLabel,
   Paper,
   Typography,
