+++ packages/frontend/src/components/SignIn/ConfirmButton.jsx
@@ -1,7 +1,6 @@
 // @flow
 import React from 'react';
 import { Button } from '@material-ui/core';
-import { Link } from 'react-router-dom';
 import type { Node } from 'react';
 
 const SignInButton = (): Node => (
