+++ packages/frontend/src/routes/index.json
@@ -1,1 +1,5 @@
-{}
+{
+  "HOME": "/",
+  "SIGNUP": "/signup",
+  "DASHBOARD": "/dashboard"
+}
